Check for bad methods in method_router
Prior to this commit, if you used a `routers.method_router`,
and it got passed a request with an HTTP method not specified
in the method_router routes, you'd get a server exception.

This commit makes it so that instead of a getting a server exception
on an unknown HTTP method, instead you get HTTP 405 Method Not Allowed.